Changed the data type of page size to short.

+++ src/Microsoft.Health.Abstractions/Data/IChangeFeedSource.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Health.Abstractions.Data
         /// <param name="pageSize">Page Size of records to fetch.</param>
         /// <param name="cancellationToken">Cancellation Token.</param>
         /// <returns>IReadOnlyCollection of T.</returns>
-        Task<IReadOnlyCollection<T>> GetRecordsAsync(long startId, int pageSize, CancellationToken cancellationToken);
+        Task<IReadOnlyCollection<T>> GetRecordsAsync(long startId, short pageSize, CancellationToken cancellationToken);
 
         /// <summary>
         /// Get resource types as a dictionary.
